Some front-end work.
Add an example of talking to the Flask + Flask-Restless backend from
Javascript.

It should be a goal to minimize the amount of template HTML whenever
possible. Instead, I'm trying to use a minimal template that we can
populate dynamically with Javascript and API. This change isn't a very
good example of that, but so far just trying to get something minimal
working. :)